+++ RELEASE_NOTES.md
@@ -15,6 +15,7 @@
 * Move documentation and code examples to the documentation website
 * Replace the local `PaginationParams` type with the `frequenz-client-common` one
 * Remove dependency to `googleapis-common-protos`
+* Replace `Energy` with `Power` for the `quantity` representation
 
 ## Bug Fixes